👌🏻 PIC-442 Fix path for court appearances within the breach endpoint
