+++ src/main/java/uk/gov/justice/probation/courtcaseservice/service/model/OffenderDetail.java
@@ -2,6 +2,7 @@ package uk.gov.justice.probation.courtcaseservice.service.model;
 
 import java.time.LocalDate;
 import java.util.List;
+import io.swagger.annotations.ApiModel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -9,6 +10,7 @@ import lombok.NoArgsConstructor;
 import uk.gov.justice.probation.courtcaseservice.controller.model.ProbationStatus;
 import uk.gov.justice.probation.courtcaseservice.restclient.communityapi.model.OtherIds;
 
+@ApiModel("The Offender detail")
 @Data
 @Builder
 @AllArgsConstructor
