+++ src/test/resources/mocks/mappings/attendances/GET_offender_attendances_server_404_X320999.json
@@ -1,7 +1,7 @@
 {
   "request": {
     "method": "GET",
-    "urlPath": "/secure/offenders/crn/X320999/convictions/2500295343/attendancesFilter",
+    "urlPath": "/secure/offenders/crn/CRNXXX/convictions/2500295343/attendancesFilter",
     "headers" : {
       "Content-Type": {
         "equalTo": "application/json"
@@ -13,6 +13,6 @@
     "headers": {
       "Content-Type": "application/json"
     },
-    "body": "{\"developerMessage\": \"Offender ID not found for CRN X320742\"}"
+    "body": "{\"developerMessage\": \"Offender ID not found for CRN CRNXXX\"}"
   }
 }
