+++ src/test/resources/mocks/mappings/offender-nsis/GET_nsi_server_error_X320741.json
@@ -1,7 +1,7 @@
 {
   "request": {
     "method": "GET",
-    "urlPath": "/secure/offenders/crn/CRNXXX/convictions/2500295343",
+    "urlPath": "/secure/offenders/crn/X320741/convictions/2500295343/nsis/500",
     "headers" : {
       "Content-Type": {
         "equalTo": "application/json"
@@ -9,10 +9,9 @@
     }
   },
   "response": {
-    "status": 404,
+    "status": 500,
     "headers": {
       "Content-Type": "application/json"
-    },
-    "body": "{\n    \"status\": 404,\n    \"developerMessage\": \"Offender with crn CRNXXX not found\"\n}"
+    }
   }
 }
